Insert ga tags in app
