+++ webpack.config.prod.js
@@ -38,6 +38,7 @@ productionWebpackConfig.plugins.push(
   new webpack.DefinePlugin({
     'process.env': {
       NODE_ENV: JSON.stringify('production'),
+      GOOGLE_ANALYTICS_ID: JSON.stringify(process.env.GOOGLE_ANALYTICS_ID),
     },
   })
 );
